Agents/customer-support-agent/pyproject.toml
[tool.poetry]
name = "customer-support-agent"
version = "0.1.0"
description = "AI Customer Support Agent with Agent Lens integration"
authors = ["Michael michael.p.dsouza@outlook.com"]
readme = "README.md"
packages = [{include = "customer_support_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
sprintlens = {path = "/Users/michaeldsouza/Documents/Wordir/AGENT_LENS/AgentVerse/Sprint_Lens_SDK", develop = true}
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.0"
pydantic = "^2.8.0"
httpx = "^0.27.0"
uvicorn = "^0.30.0"
fastapi = "^0.115.0"
requests = "^2.31.0"
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
jupyter = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
strict = true

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (requires backend connectivity)",
    "slow: marks tests as slow running",
    "unit: marks tests as unit tests"
]